Move Timecode to Timecode Holder
